Fullstack Application/backend/fastapi_app/requirements.txt
fastapi==0.115.12
llama_index==0.12.35
psycopg==3.2.6
pydantic==2.11.4
python-dotenv==1.1.0
uvicorn[standard]
psycopg[binary]
chromadb
llama-index-embeddings-huggingface
langchain
langchain-community
huggingface_hub[hf_xet]
langchain-core
langchain-community
crewai[tools]
langchain-groq
chromadb
fastapi
llama_index
numpy
opencv_python
pandas
Pillow
psycopg
pydantic
python-dotenv
segment_anything
tensorflow
torch
tf-keras
llama-index-llms-ollama
seaborn
prophet
plotly
